Renamed src module to jsonapi_query
